Step CSV logger
